producers/sql_stream_producer/requirements.txt
kafka-python
pandas
sqlalchemy
psycopg2-binary
openpyxl